requirements.txt
Cython==0.29.17
python_binance==0.7.5
numpy==1.18.1
